added next button and step for easier development
